% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_matrix.R
\name{calculate_mean_matrix}
\alias{calculate_mean_matrix}
\title{Calculate Mean Matrix of an Array}
\usage{
calculate_mean_matrix(arr)
}
\arguments{
\item{arr}{array to sum over 3rd axis.}
}
\value{
averaged matrix
}
\description{
Wrapper around sum_matrix to calculate mean matrix of an array. Runs sum_matrix with default values for index, constants and weights=TRUE.
}
\seealso{
Other sum_arrays: 
\code{\link{sum_matrix}()},
\code{\link{sum_vector}()}
}
\concept{sum_arrays}
