% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_moment.R
\name{central_moments}
\alias{central_moments}
\title{Calculate Central Moments}
\usage{
central_moments(x, exkurtosis = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{exkurtosis}{if true, calculates the ex-kurtosis instead of the kurtosis (subtract 3 from the kurtosis, to ensure the kurtosis of a Gaussian RV is 0). default True}
}
\value{
a named numeric vector consisting of the estimates for the first four centralized moments.
}
\description{
Calculate the unbiased estimates of a (univariate) sample's Mean, Variance, Skewness and Kurtosis
}
