% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_estimated_n.R
\name{compute_estimated_n_raw}
\alias{compute_estimated_n_raw}
\alias{compute_estimated_n}
\title{Estimate Degrees of Freedom by Linear Projection}
\usage{
compute_estimated_n_raw(est, theo, only_diag = FALSE)

compute_estimated_n(obj, only_diag = TRUE)
}
\arguments{
\item{est}{the empirical covariance matrix, \eqn{n^{-1}\left(Y-\hat{\mu}\right)\left(Y-\hat{\mu}\right)^{t}}}

\item{theo}{the theoretical covariance matrix}

\item{only_diag}{logical, whether to use only the diagonal (variances) or the whole matrix. Default FALSE}

\item{obj}{array of correlation matrices or a matrix of diagnolized correlation matrices.}
}
\value{
the result of the projection of theo ~ est
}
\description{
Estimate the Degrees of Freedom of a model by projecting the theoretical variance
on the estimated (empirical) variance.
When using compute_estimated_n, the theoretical variance is computed using corrmat_covariance
}
\section{Functions}{
\itemize{
\item \code{compute_estimated_n_raw}: internal function of compute_estimated_n
}}

