% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gee_variance.R
\name{compute_gee_variance}
\alias{compute_gee_variance}
\title{Compute Estimates' Variance Using a GEE Framework}
\usage{
compute_gee_variance(mod, control_arr, diagnosed_arr, est_mu = TRUE)
}
\arguments{
\item{mod}{an object created by \link[corrpops]{estimate_model}}

\item{control_arr}{array of correlation matrices of control group. can be passed as a matrix, each row is a vectorized correlation matrix}

\item{diagnosed_arr}{same as control_arr but of the diagnosed group}

\item{est_mu}{whether to use the expected value estimated in the model or the sample average. default TRUE.}
}
\value{
the Fisher information matrix of alpha
}
\description{
Estimate the estimators' variance using a Generalized Estimation Equations framework
}
\seealso{
\link[corrpops]{estimate_model} \link[corrpops]{convert_corr_array_to_data_matrix}
}
