% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_corr_array_to_data_matrix.R
\name{convert_corr_array_to_data_matrix_raw}
\alias{convert_corr_array_to_data_matrix_raw}
\alias{convert_corr_array_to_data_matrix}
\title{Convert Array of Correlation Matrices to a Data Matrix}
\usage{
convert_corr_array_to_data_matrix_raw(arr)

convert_corr_array_to_data_matrix(arr, verbose = FALSE)
}
\arguments{
\item{arr}{An array of square, symmetrical matrices or a data matrix of such form.}

\item{verbose}{should add messages?}
}
\value{
a matrix with the vectorized matrices as rows
}
\description{
Convert an array of Correlation matrices to a data matrix by vectorizing them.
Each matrix becomes a row in the output matrix. If arr is already of class matrix, nothing is perfomred.
}
\section{Functions}{
\itemize{
\item \code{convert_corr_array_to_data_matrix_raw}: internal function of convert_corr_array_to_data_matrix
}}

\seealso{
Other vectriangle: 
\code{\link{convert_data_matrix_to_corr_array_raw}()},
\code{\link{triangle2vector}()}
}
\concept{vectriangle}
