% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/corrcalc_r.R
\name{corrcalc_c}
\alias{corrcalc_c}
\alias{corrcalc_r}
\title{Formula for the Calculation of the Covariance Matrix of a Correlation Matrix}
\usage{
corrcalc_c(matr, m, order_vecti, order_vectj)

corrcalc_r(matr, m, order_vecti, order_vectj)
}
\arguments{
\item{matr}{the correlation matrix to calculate the covariance of}

\item{m}{the number of coefficients in the matrix - 0.5 p (p-1)}

\item{order_vecti}{mapping vector from i' -> i,j}

\item{order_vectj}{mapping vector from j' -> k,l}
}
\value{
the covariance matrix of the vectorized correlation matrix
}
\description{
actual calculation of the covariance parameters, Cov(R_ij, R_kl)
an internal function, shouldn't be used directly. use \link[corrpops]{corrmat_covariance}
}
\section{Functions}{
\itemize{
\item \code{corrcalc_c}: c++ application of corrcalc_r
}}

\seealso{
\link[corrpops]{triangle2vector}

Other corrcalc: 
\code{\link{corrmat_covariance_from_datamatrix}()},
\code{\link{corrmat_covariance}()}
}
\concept{corrcalc}
