% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmat_covariance.R
\name{corrmat_covariance_from_datamatrix}
\alias{corrmat_covariance_from_datamatrix}
\title{Calculate the Average Covariance Matrix of a Sample of Correlation Matrix}
\usage{
corrmat_covariance_from_datamatrix(
  datamatrix,
  fisher_z = FALSE,
  est_n = FALSE,
  only_diag = TRUE,
  nonpositive = c("ignore", "stop", "force"),
  use_cpp = TRUE,
  ncores = 1
)
}
\arguments{
\item{datamatrix}{the sample of correlation matrices. can be an array of matrices, or a matrix of vectorized correlation matrices.}

\item{fisher_z}{if true, calculate the covariance matrix of a fisher-Z transformed correlation matrix. It is assumed that the correlations are already Fisher transformed, and the matrix will under go the Inverse Fisher Transformation.}

\item{est_n}{whether to divide the covariance matrix by the estimated degrees of freedom, with a linear projection.}

\item{only_diag}{passed to \link[corrpops]{compute_estimated_n}}

\item{nonpositive}{error handling when matrix is not positive definite. can be one of 'stop', 'force', 'ignore'. if 'force' is chosen, \link[Matrix]{nearPD} will be used.}

\item{use_cpp}{whether to use c++ source code. default true.}

\item{ncores}{number of cores to use, if parallelization is in use.}
}
\value{
the averaged covariance matrix
}
\description{
Calculate the average covariance matrix of multiple correlation matrices.
}
\seealso{
Other corrcalc: 
\code{\link{corrcalc_c}()},
\code{\link{corrmat_covariance}()}
}
\concept{corrcalc}
