% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_correlation_matrices.R
\name{create_correlation_matrices}
\alias{create_correlation_matrices}
\title{Simulate a Sample of Correlation Matrices}
\usage{
create_correlation_matrices(
  n,
  df,
  Sigma,
  AR = NULL,
  MA = NULL,
  random_effect = NULL,
  ncores = 1
)
}
\arguments{
\item{n}{integer sample size.}

\item{df}{numeric parameter, “degrees of freedom”. can be lower than the dimension of Sigma}

\item{Sigma}{semi positive definite (p * p) “scale” matrix, the matrix parameter of the distribution.}

\item{AR}{a vector representing the Auto-regressive part of the multivariate ARMA process}

\item{MA}{a vector representing the Moving-average part of the multivariate ARMA process}

\item{random_effect}{generate a random effect for each matrix with \link[corrpops]{generate_random_effect_sigma}}

\item{ncores}{number of cores to use, if parallelized.}
}
\value{
an array of correlation matrices.
}
\description{
will call \link[corrpops]{rWishart_ARMA}. before returning the array, will apply \link[stats]{cov2cor} on each matrix.
}
