% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efrons_rms.R
\name{efron_bias_correction}
\alias{efron_bias_correction}
\title{Bias Correction of Efron's RMS}
\usage{
efron_bias_correction(rms, p)
}
\arguments{
\item{rms}{the RMS factor}

\item{p}{the dimension of the matrix}
}
\description{
Bias Correction of Efron's RMS
}
\seealso{
Other efron_rms: 
\code{\link{efrons_effective_sample_size}()},
\code{\link{efrons_rms_sample}()},
\code{\link{efrons_rms}()}
}
\concept{efron_rms}
