% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efrons_rms.R
\name{efrons_effective_sample_size}
\alias{efrons_effective_sample_size}
\title{Calculate Efron's Effective Sample Size}
\usage{
efrons_effective_sample_size(n, rms)
}
\arguments{
\item{n}{the number of ovservations}

\item{rms}{Efron's Root Mean Square coefficient.}
}
\description{
Estimate the Effective Sample Size, reduced due to cross-row correlation.
}
\seealso{
Other efron_rms: 
\code{\link{efron_bias_correction}()},
\code{\link{efrons_rms_sample}()},
\code{\link{efrons_rms}()}
}
\concept{efron_rms}
