% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efrons_rms.R
\name{efrons_rms}
\alias{efrons_rms}
\title{Calculate Efron's RMS Over A Correlation Matrix}
\usage{
efrons_rms(m, p = NULL)
}
\arguments{
\item{m}{a correlation matrix}

\item{p}{the dimension of the matrix. defult null. if not null, conduct a bias correction.}
}
\description{
Calculate Efron's RMS Over A Correlation Matrix
}
\references{
Bradley Efron. Large-scale inference: empirical Bayes methods for estimation, testing, and prediction, volume 1. Cambridge University Press, 2012.
}
\seealso{
Other efron_rms: 
\code{\link{efron_bias_correction}()},
\code{\link{efrons_effective_sample_size}()},
\code{\link{efrons_rms_sample}()}
}
\concept{efron_rms}
