% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efrons_rms.R
\name{efrons_rms_sample}
\alias{efrons_rms_sample}
\title{Calculate the Average Efron's RMS
Calculate the average Efron's RMS over a sample of corraltion matrices}
\usage{
efrons_rms_sample(arr, p = NULL)
}
\arguments{
\item{arr}{an array of correlation matrices (can be in data matrix-form)}

\item{p}{the dimension of the matrix}
}
\description{
Calculate the Average Efron's RMS
Calculate the average Efron's RMS over a sample of corraltion matrices
}
\seealso{
Other efron_rms: 
\code{\link{efron_bias_correction}()},
\code{\link{efrons_effective_sample_size}()},
\code{\link{efrons_rms}()}
}
\concept{efron_rms}
