% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate Model}
\usage{
estimate_model(
  control_arr,
  diagnosed_arr,
  dim_alpha = 1,
  LinkFunc = LinkFunctions$multiplicative_identity,
  model_reg_config = list(),
  matrix_reg_config = list(),
  iid_config = list(),
  cov_config = list(),
  raw_start = FALSE,
  bias_correction = FALSE,
  early_stop = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{control_arr}{array of control group correlation matrices. either an array or data matrix form}

\item{diagnosed_arr}{array of diagnosed group correlation matrices. either an array or data matrix form}

\item{dim_alpha}{the number of columns in alpha. default 1}

\item{LinkFunc}{a list of function. must include func, inverse, rev_func and null_value. see \link[corrpops]{LinkFuncSkeleton}}

\item{model_reg_config}{see \link[corrpops]{configurations}. arguments passed will override the defaults.}

\item{matrix_reg_config}{see \link[corrpops]{configurations}. arguments passed will override the defaults.}

\item{iid_config}{list of two lists named 'iter_config' and 'optim_config', for the optimization of the model with identity matrix covariance matrix. see \link[corrpops]{configurations}. arguments passed will override the defaults.}

\item{cov_config}{list of two lists named 'iter_config' and 'optim_config', for the optimization of the model with a specified covariance matrix. see \link[corrpops]{configurations}. arguments passed will override the defaults.}

\item{raw_start}{if true, don't optimize with the identity matrix before optimizing with a specified covariance matrix}

\item{bias_correction}{if true, correct the estimates to the median: a' = a - med(a) + null_value}

\item{early_stop}{if true, stop the optimization of the joint loss function (of theta and alpha) didn't decrease in the last iteration.}

\item{verbose}{if true, print status to console}
}
\value{
a list of the following:

\itemize{
\item theta: a matrix, the estimates of theta
}

\itemize{
\item alpha: a matrix, the estimates of alpha
}

\itemize{
\item LinkFunc: the link function used same as the parameter LinkFunc
}

\itemize{
\item regularization: the same as model_reg_config parameter
}

\itemize{
\item vcov: the weighting matrix used in the optimization - the inverse of the correlations' covariance matrix
}

\itemize{
\item convergence: a vector with the convergence in each iteration. see \link[stats]{optim}
}

\itemize{
\item steps: the estimates of theta, alpha in each iteration
}

\itemize{
\item log_optim: if optim_config$log_optim=TRUE, will return the output of \link[stats]{optim} for each iteration. else, NA.
}
}
\description{
todo: explain
}
\seealso{
\link[corrpops]{configurations}
}
