% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_jacknife.R
\name{estimate_model_jacknife}
\alias{estimate_model_jacknife}
\title{Jackknife Estimation of Model}
\usage{
estimate_model_jacknife(
  control_arr,
  diagnosed_arr,
  dim_alpha = 1,
  alpha0 = NULL,
  theta0 = NULL,
  LinkFunc = LinkFunctions$multiplicative_identity,
  model_reg_config = list(),
  matrix_reg_config = list(),
  iid_config = list(iter_config = list(minit = 0)),
  cov_config = list(),
  return_gee = FALSE,
  jack_control = TRUE,
  bias_correction = FALSE,
  early_stop = FALSE,
  verbose = TRUE,
  ncores = 1
)
}
\arguments{
\item{control_arr}{array of control group correlation matrices. either an array or data matrix form}

\item{diagnosed_arr}{array of diagnosed group correlation matrices. either an array or data matrix form}

\item{dim_alpha}{the number of columns in alpha. default 1}

\item{alpha0}{starting point for alpha in the optimization. if null (the default), will use LinkFunc$null_value}

\item{theta0}{starting point for alpha in the optimization. if null (the default), will the average matrix of all subjects}

\item{LinkFunc}{a list of function. must include func, inverse, rev_func and null_value. see \link[corrpops]{LinkFuncSkeleton}}

\item{model_reg_config}{see \link[corrpops]{configurations}. arguments passed will override the defaults.}

\item{matrix_reg_config}{see \link[corrpops]{configurations}. arguments passed will override the defaults.}

\item{iid_config}{list of two lists named 'iter_config' and 'optim_config', for the optimization of the model with identity matrix covariance matrix. see \link[corrpops]{configurations}. arguments passed will override the defaults.}

\item{cov_config}{list of two lists named 'iter_config' and 'optim_config', for the optimization of the model with a specified covariance matrix. see \link[corrpops]{configurations}. arguments passed will override the defaults.}

\item{return_gee}{if true, calculate the gee estimate of variance in each jackknife}

\item{jack_control}{if false, don't jackknife control subjects}

\item{bias_correction}{if true, correct the estimates to the median: a' = a - med(a) + null_value}

\item{early_stop}{if true, stop the optimization of the joint loss function (of theta and alpha) didn't decrease in the last iteration.}

\item{verbose}{if true, print status to console}

\item{ncores}{number of cores to use in parallel}
}
\value{
a list of the following:

\itemize{
\item theta: a matrix, the estimates of theta from each jackknife iteration vectorized
}

\itemize{
\item alpha: a matrix, the estimates of alpha from each jackknife iteration vectorized
}

\itemize{
\item gee_var: if return_gee=TRUE, an array of the gee estimate of variance from each jackknife iteration
}

\itemize{
\item convergence: a matrix consisting of vectors with the convergence in each iteration. see \link[stats]{optim}
}

\itemize{
\item is_diagnosed: a vector with 0,1 indicating of a control or diagnosed subjet was omitted in the iteration
}

\itemize{
\item LinkFunc: the link function used same as the parameter LinkFunc
}

\itemize{
\item regularization: the same as model_reg_config parameter
}
}
\description{
todo: explain
}
\seealso{
\link[corrpops]{estimate_model}
}
