% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{force_symmetry}
\alias{force_symmetry}
\title{Force a Matrix to be Symmetrical}
\usage{
force_symmetry(matr)
}
\arguments{
\item{matr}{the matrix to enforce symmetry}
}
\description{
Force a matrix to be symmetrical by averaging between the matrix and it's transpose
}
