% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwishart_extension.R
\name{generate_random_effect_sigma}
\alias{generate_random_effect_sigma}
\title{Generate a Covariance Matrix with a Random Effect}
\usage{
generate_random_effect_sigma(Sigma, random_effect = NULL)
}
\arguments{
\item{Sigma}{Expected value of the covariance matrix}

\item{random_effect}{the control on the random effect size. for 0 < random_effect <= Inf, a covariance matrix with distribution Wishart(Sigma, q)/q is generated with q = p * (1 + 1/random_effect), so a larger random_effect size means a larger variance in the results.}
}
\value{
a covariance matrix with Expected value Sigma and variance roughly proportional to random_effect
}
\description{
Generate a Covariance matrix with expected value Sigma, and a random effect controlled by random_effect parameter
}
\seealso{
Other rWishart2: 
\code{\link{rWishart2}()},
\code{\link{rWishart_ARMA}()}
}
\concept{rWishart2}
