% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_invertable_arma.R
\name{is_invertable_arma}
\alias{is_invertable_arma}
\title{Check Invertibility/Stationarity of an AR/MA Process}
\usage{
is_invertable_arma(coefs, tol = 0.001)
}
\arguments{
\item{coefs}{a numeric vector with the AR (or MA) coefficients}

\item{tol}{the tolerance search for the roots.}
}
\value{
true if the process is invertible/stationary.
}
\description{
Check if an AR process is stationary or an MA process is invertible,
by testing if the polynomial defined by it's coefficients doesn't have a root between -1 and 1
}
