% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\docType{data}
\name{LinkFuncSkeleton}
\alias{LinkFuncSkeleton}
\title{Skeleton of Link Function}
\format{
An object of class \code{list} of length 5.
}
\usage{
LinkFuncSkeleton
}
\description{
Skeleton of link function consumed by \link[corrpops]{estimate_model}.
has the following elements:
\itemize{
\item name: optional, a string specifing the link function's name
\item func: mandatory, a function with input t (theta), a vectorized correlation matrix, a (alpha) and d, the number of columns in alpha. the output would be a correlation matrix with alpha effect
\item inverse: optinal, a function with input a, that inverts alpha. for example, of we specify func to calculate exp(a), inverse would be log(a)
\item rev_func: mandatory, a function with input datamatrix, a vectorized array of correlation matrices, a (alpha) and d, the number of columns in alpha. the output would be the array with the alpha effect reversed
\item null_value: mandatory, the null value of the link function
}
}
\keyword{datasets}
