% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\docType{data}
\name{multiplicative_identity}
\alias{multiplicative_identity}
\alias{additive_quotent}
\alias{LinkFunctions}
\title{Default Link Functions for model}
\format{
An object of class \code{list} of length 5.

An object of class \code{list} of length 5.

An object of class \code{list} of length 2.
}
\usage{
multiplicative_identity

additive_quotent

LinkFunctions
}
\description{
Default Link Functions for model
}
\section{Functions}{
\itemize{
\item \code{multiplicative_identity}: The multiplicative identity link function

\item \code{additive_quotent}: The additive quotent link function
}}

\seealso{
\link[corrpops]{LinkFuncSkeleton}
}
\keyword{datasets}
