% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_power}
\alias{matrix_power}
\title{Raise a Square Matrix to a Non-Integer Power}
\usage{
matrix_power(x, power)
}
\arguments{
\item{x}{a square matrix}

\item{power}{a real}
}
\value{
the matrix raised to the power
}
\description{
Raise a Square Matrix to a non-integer power using it's Spectral Decomposition
}
