% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_sqrt_diag}
\alias{mean_sqrt_diag}
\title{Get the Average of the Square-rooted Diagonal of a Matrix}
\usage{
mean_sqrt_diag(x)
}
\arguments{
\item{x}{a square matrix}
}
\description{
Get the Average of the Square-rooted Diagonal of a Matrix
}
