% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_utils.R
\name{optimiser}
\alias{optimiser}
\title{Iterative Call for optim}
\usage{
optimiser(
  control_datamatrix,
  diagnosed_datamatrix,
  alpha0,
  theta0,
  weight_matrix,
  dim_alpha,
  LinkFunc,
  model_reg_config,
  matrix_reg_config,
  iter_config,
  optim_config,
  early_stop,
  verbose
)
}
\arguments{
\item{control_datamatrix}{data matrix of control subjects}

\item{diagnosed_datamatrix}{data matrix of diagnosed subjects}

\item{alpha0}{starting point for alpha}

\item{theta0}{starting point for theta}

\item{weight_matrix}{the weighting matrix to be used in the mahalonobis sum of squares}

\item{dim_alpha}{dimension of alpha}

\item{LinkFunc}{\link[corrpops]{LinkFuncSkeleton}}

\item{model_reg_config}{see \link[corrpops]{configurations}}

\item{matrix_reg_config}{see \link[corrpops]{configurations}}

\item{iter_config}{see \link[corrpops]{configurations}}

\item{optim_config}{see \link[corrpops]{configurations}}

\item{early_stop}{if true, stop the optimization of the joint loss function (of theta and alpha) didn't decrease in the last iteration.}

\item{verbose}{if true, print status to console}
}
\value{
see \link[corrpops]{estimate_model}
}
\description{
Iterative Call for optim
}
\seealso{
\link[corrpops]{estimate_model}
}
