% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize_matrix.R
\name{regularize_matrix}
\alias{regularize_matrix}
\title{Regularize Matrix}
\usage{
regularize_matrix(
  matr,
  const = 1,
  method = c("constant", "avg_diag", "increase_diag"),
  only_if_singular = TRUE
)
}
\arguments{
\item{matr}{the matrix to be regularized}

\item{const}{the constant to be used in the regularization}

\item{method}{how to regularize the matrix. can be one of 'constant', 'avg_diag', 'increase_diag'. see details for more information.}

\item{only_if_singular}{if true, will regularize only if the matrix is singular}
}
\value{
if method = 'constant', will add the constant to the diagonal. if method = 'avg_diag', will calculate a weighted average of the original matrix and an identity matrix multiplied by the avregae value of the diagonal, with weights 1 - const and const. if method = 'avg_diag', will calculate a weighted average of the original matrix and the diagonal of the matrix, with weights 1 - const and const
}
\description{
Regularize a square matrix by manipulating it's diagonal
}
