% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_zeros}
\alias{remove_zeros}
\title{Replace a Value in a vector/array}
\usage{
remove_zeros(x, rem = 0, rep = 1)
}
\arguments{
\item{x}{the array}

\item{rem}{the value to replace, default 0}

\item{rep}{the value to replace with, default 1}
}
\value{
the array with rem replaces with rep
}
\description{
Replace a Value in a vector/array
}
