% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwishart_extension.R
\name{rWishart2}
\alias{rWishart2}
\title{Random Wishart Distributed Matrices with df < p}
\usage{
rWishart2(n, df, Sigma, random_effect = NULL)
}
\arguments{
\item{n}{integer sample size.}

\item{df}{numeric parameter, “degrees of freedom”. can be lower than the dimension of Sigma}

\item{Sigma}{semi positive definite (p * p) “scale” matrix, the matrix parameter of the distribution.}

\item{random_effect}{generate a random effect for each matrix with \link[corrpops]{generate_random_effect_sigma}}
}
\value{
a numeric array, of dimension p * p * n, where each matrix is semi positive definite covariance matrix, a realization of the (possibly degenerate) Wishart distribution W_p(Sigma, df), where Sigma is possibly an RV itself.
}
\description{
Generate n random matrices, distributed according to the (possibly degenerate) Wishart distribution with parameters Sigma and df, W_p(Sigma, df).
with df possibly lower than p, Sigma can be semi positive definite and with a random effect.
if df >= p, random_effect is null and Sigma is positive definite, will call \link[stats]{rWishart}.
}
\seealso{
\link[stats]{rWishart}

Other rWishart2: 
\code{\link{generate_random_effect_sigma}()},
\code{\link{rWishart_ARMA}()}
}
\concept{rWishart2}
