% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwishart_extension.R
\name{rWishart_ARMA}
\alias{rWishart_ARMA}
\title{Random Covariance Matrices of Correlated Multivariate Normal RVs}
\usage{
rWishart_ARMA(
  n,
  df,
  Sigma,
  random_effect = NULL,
  AR = NULL,
  MA = NULL,
  ncores = 1
)
}
\arguments{
\item{n}{integer sample size.}

\item{df}{numeric parameter, “degrees of freedom”. can be lower than the dimension of Sigma}

\item{Sigma}{semi positive definite (p * p) “scale” matrix, the matrix parameter of the distribution.}

\item{random_effect}{generate a random effect for each matrix with \link[corrpops]{generate_random_effect_sigma}}

\item{AR}{a vector representing the Auto-regressive part of the multivariate ARMA process}

\item{MA}{a vector representing the Moving-average part of the multivariate ARMA process}

\item{ncores}{number of cores to use, if parallelized.}
}
\value{
a numeric array, of dimension p * p * n, where each matrix is semi positive definite covariance matrix, a realization proportional to the (possibly degenerate) Wishart distribution W_p(Sigma, df), where Sigma is possibly an RV itself.
}
\description{
Generate n random matrices, distributed proportionally to the (possibly degenerate) Wishart distribution with parameters Sigma and df, W_p(Sigma, df),
with df able to be lower than p, Sigma can be semi positive definite and a random effect.
Also, unlike the underlying assumption of the Wishart distribution, the covariance matrices are simulated using correlated multivariate Gaussian RVs - not i.i.d.
if AR & MA are null, will call \link[corrpops]{rWishart2}
}
\seealso{
\link[stats]{rWishart}

Other rWishart2: 
\code{\link{generate_random_effect_sigma}()},
\code{\link{rWishart2}()}
}
\concept{rWishart2}
