% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_matrix.R
\name{sum_matrix}
\alias{sum_matrix}
\title{Sum Over Array's Matrices}
\usage{
sum_matrix(arr, index, constants, weights = FALSE)
}
\arguments{
\item{arr}{array to sum over it's 3rd axis.}

\item{index}{index of matrices to sum. default is to sum over all matrices.}

\item{constants}{constants to multiply the matrices before summation. default is 1 for all constants}

\item{weights}{should treat constants as weights (sum of all constants = 1)? default false}
}
\value{
a matrix, built from the (possible weighted) sum of specified matrices
}
\description{
Sum over specified Array matrices, possibly weighted.
}
\seealso{
Other sum_arrays: 
\code{\link{calculate_mean_matrix}()},
\code{\link{sum_vector}()}
}
\concept{sum_arrays}
