% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_utils.R
\name{sum_of_squares}
\alias{sum_of_squares}
\title{Loss Function for Optimization Process on Alpha}
\usage{
sum_of_squares(
  alpha,
  theta,
  diagnosed_datamatrix,
  inv_sigma,
  LinkFunc,
  sigma,
  dim_alpha = 1,
  reg_lambda = 0,
  reg_p = 2
)
}
\arguments{
\item{alpha}{estimate of alpha}

\item{theta}{estimate of theta}

\item{diagnosed_datamatrix}{data matrix of diagnosed subjects}

\item{inv_sigma}{covariance matrix of correlation data matrix, inverted}

\item{LinkFunc}{\link[corrpops]{LinkFuncSkeleton}}

\item{sigma}{if inv_sigma is missing, invert this matrix and use it as inv_sigma}

\item{dim_alpha}{dimension of alpha}

\item{reg_lambda}{see model_reg in \link[corrpops]{configurations}}

\item{reg_p}{see model_reg in \link[corrpops]{configurations}}
}
\value{
value of sum of squares
}
\description{
Loss Function for Optimization Process on Alpha
}
