% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_matrix.R
\name{sum_vector}
\alias{sum_vector}
\title{Sum Over Matrix Rows / Columns}
\usage{
sum_vector(matr, index, constants, weights = FALSE, by_row = TRUE)
}
\arguments{
\item{matr}{matrix to sum over rows / columns}

\item{index}{index of rows / columns to sum. default is to sum over all rows.}

\item{constants}{constants to multiply the indices before summation. default is 1 for all constants}

\item{weights}{should treat constants as weights? if true, weights <- weights/sum(weights). default false}

\item{by_row}{should sum over rows (default) or by columns?}
}
\value{
a vector, built from the (possible weighted) sum of specified vectors
}
\description{
Sum over specified matrix rows / columns, possibly weighted. Useful for creating multivariate ARMA processes.
}
\seealso{
Other sum_arrays: 
\code{\link{calculate_mean_matrix}()},
\code{\link{sum_matrix}()}
}
\concept{sum_arrays}
