% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_corr_mat.R
\name{test_corr_mat}
\alias{test_corr_mat}
\title{Test if all Matrices of an Array are Positive Definite and without Missing Values}
\usage{
test_corr_mat(arr)
}
\arguments{
\item{arr}{an array of square matrices}
}
\value{
a list of index vectors:

\itemize{
\item which_not_positive_definite: which matrices in the array aren't positive_definite
}

\itemize{
\item which_na: which matrices in the array have missing values
}
}
\description{
Test if all Matrices of an Array are Positive Definite and without Missing Values
}
