% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tga_arrays}
\alias{tga_arrays}
\title{Correlation Matrices from the Research of Michael Peer et al on the Functional Connectivity of TGA Patients.}
\format{
A list with 2 86x86 arrays:
\describe{
\item{control}{control subjects with n = 23}
\item{diagnosed}{diagnosed subjects with n = 14}
}
}
\usage{
tga_arrays
}
\description{
A dataset containing 86x86 correlation matrices of 23 control subjects and 14 TGA-diagnosed patients.
}
\references{
Michael Peer et al. Reversible functional connectivity disturbances during transient global amnesia. Annals of neurology, 75(5):634–643, 2014.
}
\keyword{datasets}
