% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_utils.R
\name{theta_of_alpha}
\alias{theta_of_alpha}
\title{Compute Theta as a Function of Alpha}
\usage{
theta_of_alpha(
  alpha,
  control_datamatrix,
  diagnosed_datamatrix,
  LinkFunc,
  d = 1
)
}
\arguments{
\item{alpha}{estimate of alpha}

\item{control_datamatrix}{data matrix of control subjects}

\item{diagnosed_datamatrix}{data matrix of diagnosed subjects}

\item{LinkFunc}{\link[corrpops]{LinkFuncSkeleton}}

\item{d}{dimension of alpha}
}
\value{
estimate of theta based on control and doagnosed subjects
}
\description{
Compute Theta as a Function of Alpha
}
