% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle2vector.R
\name{triangle2vector}
\alias{triangle2vector}
\alias{vector2triangle}
\title{Transform a Symmetric Matrix to it's Vectorised Form and Vise Versa}
\usage{
triangle2vector(matr, diag = FALSE)

vector2triangle(vect, diag = FALSE, diag_value = NA)
}
\arguments{
\item{matr}{the matrix to vectorize it's lower triangle}

\item{diag}{if true, vectorize (or construct) the diagonal of the matrix as well}

\item{vect}{the vector to transform to a symmetric matrix}

\item{diag_value}{if diag is set to false, what value to place on the diagonal. default is NA, for correlation matrices set to 1.}
}
\value{
the transformed matrix / vector
}
\description{
A symmetric matrix can be entirely summarized with one of its triangles.
these functions transform a matrix's lower triangle to a vectorized form, and transform a vector to a symmetric matrix.
In correlation matrices, the diagonal is redundant, so it is possible to vectorize the off diagonal triangle.
}
\section{Functions}{
\itemize{
\item \code{vector2triangle}: inverse of triangle2vector
}}

\seealso{
Other vectriangle: 
\code{\link{convert_corr_array_to_data_matrix_raw}()},
\code{\link{convert_data_matrix_to_corr_array_raw}()}

Other vectriangle: 
\code{\link{convert_corr_array_to_data_matrix_raw}()},
\code{\link{convert_data_matrix_to_corr_array_raw}()}
}
\concept{vectriangle}
