% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vect_norm}
\alias{vect_norm}
\title{Calculate the Mahalonobis Norm of a Vector}
\usage{
vect_norm(x, matr = NULL, sqrt = TRUE, solve_matr = FALSE)
}
\arguments{
\item{x}{a vector}

\item{matr}{the weighting matrix in the Mahalonobis norm. if null, the l2 norm is calculated.}

\item{sqrt}{if false, don't take the sqrt of the product}

\item{solve_matr}{if true, invert matr before calculating the norm}
}
\value{
the norm, a real number >= 0
}
\description{
Calculate the Mahalonobis Norm of a Vector
}
